+++ sample/build.gradle.kts
@@ -40,7 +40,6 @@ dependencies {
 
     implementation("androidx.compose.ui:ui:$composeVersion")
     implementation("androidx.compose.material:material:$composeVersion")
-    implementation("androidx.compose.material:material:$composeVersion")
     implementation("androidx.activity:activity-compose:1.4.0")
 
     implementation("androidx.core:core-ktx:1.7.0")
